The button 'Start Game' is created and added functionality to it
