+++ src/routes/Game/index.js
@@ -1,0 +1,11 @@
+import './style.module.css';
+
+const GamePage = () => {
+  return (
+    <div>
+      This is Game Page!!!
+    </div>
+  );
+};
+
+export default GamePage;
